Product validation

+++ src/main/java/com/alex/orders/service/OrderService.java
@@ -1,10 +1,11 @@
 package com.alex.orders.service;
 
 import com.alex.orders.entity.Order;
+import com.alex.orders.entity.OrderProducts;
+import com.alex.orders.exceptions.OrderProductAlreadyExistsException;
+import com.alex.orders.repository.OrderProductsRepo;
 import com.alex.orders.repository.OrderRepo;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,34 +15,40 @@ public class OrderService {
   @Autowired
   OrderRepo orderRepo;
 
-  public List<Order> getAllOrder() {
-    List<Order> ordersList = orderRepo.findAll();
+  @Autowired
+  OrderProductsRepo orderProductsRepo;
 
-    if (ordersList.size() > 0) {
-      return ordersList;
-    } else {
-      return new ArrayList<Order>();
-    }
+  public Order getOrderById(Long id) {
+    return orderRepo.findById(id).orElse(null);
   }
 
-  public Optional<Order> getOrderById(Long id) {
-    return orderRepo.findById(id);
+  public List<Order> getAllOrders() {
+    return orderRepo.findAll();
   }
 
-  //maybe invoke the external validation here when creating an order
-  //check in repo if there already exists an order with the same list of products
-  //swagger by openapi? Mock api? mockapi.io
   public Order createOrder(Order order) {
-    Order newOrder = new Order();
-    newOrder.setProductDescription(order.getProductDescription());
-    newOrder.setOrderPrice(order.getOrderPrice());
-    newOrder.setProducts(order.getProducts()); //compare this list with other orders in repo
-    newOrder = orderRepo.save(newOrder);
+    if (orderRepo.existsById(order.getOrderId())) {
+      // Check if any of the orderProducts already exist in the database.
+      for (OrderProducts orderProduct : order.getOrderProducts()) {
+        if (
+          orderProduct.getId() != null &&
+          orderProductsRepo.existsById(orderProduct.getId())
+        ) {
+          throw new OrderProductAlreadyExistsException(
+            "OrderProduct with ID " + orderProduct.getId() + " already exists."
+          );
+        }
+      }
+    }
 
-    return newOrder;
+    return orderRepo.save(order);
   }
 
-  public void deleteOrder(Long id) {
-    orderRepo.deleteById(id);
+  public boolean deleteOrderById(Long id) {
+    if (orderRepo.existsById(id)) {
+      orderRepo.deleteById(id);
+      return true;
+    }
+    return false;
   }
 }
